+++ app/src/main/java/com/vuerts/weather/domain/forecast/model/units/MilliBars.kt
@@ -1,4 +1,4 @@
 package com.vuerts.weather.domain.forecast.model.units
 
 @JvmInline
-value class MilliBars(val millibars: Long)
+value class MilliBars(val millibars: Double)
